trim white space and add Apache Liscence and modify test unit

+++ hbase-server/src/main/java/org/apache/hadoop/hbase/regionserver/MetricsHeapMemoryManagerWrapperImpl.java
@@ -1,3 +1,21 @@
+/*
+ * Copyright The Apache Software Foundation
+ *
+ * Licensed to the Apache Software Foundation (ASF) under one or more
+ * contributor license agreements. See the NOTICE file distributed with this
+ * work for additional information regarding copyright ownership. The ASF
+ * licenses this file to you under the Apache License, Version 2.0 (the
+ * "License"); you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
+ * WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
+ * License for the specific language governing permissions and limitations
+ * under the License.
+ */
 package org.apache.hadoop.hbase.regionserver;
 
 import java.lang.management.ManagementFactory;
@@ -11,13 +29,13 @@ import org.apache.hadoop.hbase.io.hfile.ResizableBlockCache;
 @InterfaceAudience.Private
 public class MetricsHeapMemoryManagerWrapperImpl
     implements MetricsHeapMemoryManagerWrapper {
-  
+
   private long maxHeapSize = ManagementFactory.getMemoryMXBean().getHeapMemoryUsage().getMax();
-  
+
   private final HRegionServer server;
   private final RegionServerAccounting rsAccounting;
   private final ResizableBlockCache blockCache;
-  
+
   public MetricsHeapMemoryManagerWrapperImpl(final Server server,
       final RegionServerAccounting rsAccounting, ResizableBlockCache blockCache) {
     this.server = (HRegionServer) server;
@@ -57,7 +75,7 @@ public class MetricsHeapMemoryManagerWrapperImpl
 
   @Override
   public float getBlockCacheUsed() {
-    return (float) (getBlockCacheUsedSize() / maxHeapSize);
+    return (float) getBlockCacheUsedSize() / (float) maxHeapSize;
   }
 
   @Override
@@ -67,14 +85,14 @@ public class MetricsHeapMemoryManagerWrapperImpl
 
   @Override
   public float getMemStoreUsed() {
-    return (float) (getMemStoreUsedSize() / maxHeapSize);
+    return (float) getMemStoreUsedSize() / (float) maxHeapSize;
   }
 
   @Override
   public long getMemStoreUsedSize() {
     return rsAccounting.getGlobalMemstoreSize();
   }
-  
+
   private MemoryUsage getMemoryUsage() {
     return ManagementFactory.getMemoryMXBean().getHeapMemoryUsage();
   }
